+++ judge-impl/src/main/java/com/dwarfeng/judge/impl/dao/preset/SinkerSupportPresetCriteriaMaker.java
@@ -1,18 +1,46 @@
 package com.dwarfeng.judge.impl.dao.preset;
 
+import com.dwarfeng.judge.stack.service.SinkerSupportMaintainService;
 import com.dwarfeng.subgrade.sdk.hibernate.criteria.PresetCriteriaMaker;
 import org.hibernate.criterion.DetachedCriteria;
+import org.hibernate.criterion.MatchMode;
+import org.hibernate.criterion.Restrictions;
 import org.springframework.stereotype.Component;
+
+import java.util.Arrays;
 
 @Component
 public class SinkerSupportPresetCriteriaMaker implements PresetCriteriaMaker {
 
-    @SuppressWarnings("SwitchStatementWithTooFewBranches")
     @Override
     public void makeCriteria(DetachedCriteria criteria, String preset, Object[] objs) {
         switch (preset) {
+            case SinkerSupportMaintainService.ID_LIKE:
+                idLike(criteria, objs);
+                break;
+            case SinkerSupportMaintainService.LABEL_LIKE:
+                labelLike(criteria, objs);
+                break;
             default:
                 throw new IllegalArgumentException("无法识别的预设: " + preset);
+        }
+    }
+
+    private void idLike(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            String pattern = (String) objects[0];
+            detachedCriteria.add(Restrictions.like("stringId", pattern, MatchMode.ANYWHERE));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
+        }
+    }
+
+    private void labelLike(DetachedCriteria detachedCriteria, Object[] objects) {
+        try {
+            String pattern = (String) objects[0];
+            detachedCriteria.add(Restrictions.like("label", pattern, MatchMode.ANYWHERE));
+        } catch (Exception e) {
+            throw new IllegalArgumentException("非法的参数:" + Arrays.toString(objects));
         }
     }
 }